v1/admin/requirements.txt
streamlit>=1.40
psycopg[binary]>=3.2
paramiko>=3.5
bcrypt>=4.2
pyyaml>=6.0
httpx>=0.28
